Fix i18n authorbox_name string error
This is due to an upstream issue in Hugo ≥0.76 (Upgrade to go-i18n v2).
Three months have passed, so I have to look for workarounds on our side.
The most popular option (replace `.Count` to `.`) doesn't suit us
because it breaks backward compatibility.

This fix will allow us to preserve backward compatibility with older
versions.

Fix #237